Working

+++ CampusCloudUniversityCampusSystem/src/main/java/com/mayank/CampusCloudUniversityCampusSystem/repository/AdminRepo.java
@@ -3,5 +3,5 @@ package com.mayank.CampusCloudUniversityCampusSystem.repository;
 import com.mayank.CampusCloudUniversityCampusSystem.model.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepo extends JpaRepository<Admin,Integer> {
+public interface AdminRepo extends JpaRepository<Admin,Long> {
 }
