+++ CampusCloudUniversityCampusSystem/src/main/java/com/mayank/CampusCloudUniversityCampusSystem/repository/FacultyRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FacultyRepo extends JpaRepository<Faculty,Integer> {
+public interface FacultyRepo extends JpaRepository<Faculty,Long> {
 
     Optional<Faculty>  findFacultyByUnivId(String univId);
