+++ CampusCloudUniversityCampusSystem/src/main/java/com/mayank/CampusCloudUniversityCampusSystem/repository/StudentRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StudentRepo extends JpaRepository<Student,Integer> {
+public interface StudentRepo extends JpaRepository<Student,Long> {
 
     Optional<Student> findByRollNo(String rollNo);
